Corrected Prep Maps
Fixed the issue with the prep maps becoming unresponsive. Each map is
now on its own slide again. The problem was a clash between the api
required for the js to run which is at the bottom of the trail html
page, and one of the google api commands in the code Eric had for the
trail map to load. I replaced all the code Eric had at the top of the
trail html page so that the trail map loads and runs with the js api
key.